Working but not entierly finished implementation of the ?api built-in. Excerpt from the version history:
New built-ins: ?api and ?has_api. This exposes the Java API of the object, if the configuration allows it. This is mostly meant to be used when a Map or List (or other Collection) has some custom methods (business logic methods) that you need to call from the template. As FreeMarker (by default) exposes these as clean FTL hash and sequence values, you couldn't access the custom methods. For example, myMap.myMethod() in the template basically translates to ((Method) myMap.get("myMethod")).invoke(...) in Java. Now you can write myMap?api.myMethod(), which basically translates to myMap.myMethod() in Java. Using ?api also happens to offer a workaround for the lack of non-String Map key support in FTL's [] operator (as in myMap[key]), because now you can write myMap?api.get(nonStringKey).

However, ?api is not enabled by default, for now at least. The api_builtin_enabled setting has to be set to true for that. Also the object_wrapper and the value itself has to support ?api. DefaultObjectWrapper with incompatibleImprovements 2.3.22 (or with useAdaptersForContainers set to true) and pure BeansWrapper supports it, for Map-s, List-s and other Collection-s.