Implement a #ftl onlyIncludeOnce setting.
Using this setting allows files to be only included and interpreted once.

This is the freemarker equivalent of MS C++ [#pragma once]. Which was a fix for
the common C-header pattern:

[#ifndef MYFILE_H]
[#define MYFILE_H]

... rest of myfile.h

[#endif]

This looks in freemarker like that:

<#if !(MYFILE_INCLUDED!false>
<#assign MYFILE_INCLUDED=true>

... rest of myfile
</#if>

which of course is also very ugly.

As a freemarker project gets bigger and more complex many lib templates only
containing macros and functions arise. And the same old C++ problems with it:
Those libs depend on each other, some include the required other libs explicit,
some just assume they have already be included. Which of course
causes problems from time to time.

Using <#ftl onlyIncludeOnce=true> allows to always explicit specify the
dependencies without getting a runtime penalty or some unwanted side effect
(e.g. reset of global variables) from it.